Add Linux Support

+++ Assets/Forge/Scripts/Editor/Importers/LevelImporterWindow.cs
@@ -763,6 +763,13 @@ public class LevelImporterWindow : EditorWindow
             //PackerImporterWindow.Import(assetImports, true);
             //return;
 
+            // We've decided that all custom maps for UYA must support both NTSC & PAL
+            // So to that end in order to import a UYA map (base map) you must also have the PAL iso
+            if (ImportSourceIsUYA() && string.IsNullOrEmpty(forgeSettings.PathToCleanUyaPalIso)) throw new Exception("Missing UYA PAL iso");
+
+            // save open scenes
+            EditorSceneManager.SaveOpenScenes();
+
             // clear map assets folder on fresh import
             if (Directory.Exists(destMapFolder)) Directory.Delete(destMapFolder, true);
             if (Directory.Exists(tempPalBinFolder)) Directory.Delete(tempPalBinFolder, true);
@@ -825,7 +832,7 @@ public class LevelImporterWindow : EditorWindow
                 var palWadFile = Path.Combine(tempPalBinFolder, "pal.wad");
                 ExtractWadFromISO(forgeSettings.PathToCleanUyaPalIso, GetLevelId(), palWadFile);
                 if (!DecompressAndUnpackLevelCodeFromWad(palWadFile)) return;
-                
+
                 // import PAL code
                 ImportCode(tempPalBinFolder, destMapFolder, GameRegion.PAL, assetImports, rootGo);
             }
@@ -902,6 +909,9 @@ public class LevelImporterWindow : EditorWindow
             var assets = GameObject.FindObjectsOfType<MonoBehaviour>().Where(x => x is IAsset).Select(x => x as IAsset);
             foreach (var asset in assets)
                 asset.UpdateAsset();
+                
+            // save new scene
+            EditorSceneManager.SaveOpenScenes();
         }
         catch (Exception ex)
         {
@@ -1226,12 +1236,13 @@ public class LevelImporterWindow : EditorWindow
         var levelId = GetLevelId();
         var bResult = false;
 
-        var unpackSounds = importMobys > 0;
-        var unpackAssets = importMobys > 0 || importSky > 0 || importCollision > 0 || importTfrags > 0 || importTies > 0 || importShrubs > 0;
-        var unpackGameplay = importMobys > 0 || importWorldConfig > 0 || importMisc > 0;
-        var unpackOcclusion = importTies > 0 || importTfrags > 0 || importMobys > 0;
-        var unpackWorldInstances = importTies > 0 || importShrubs > 0 || importWorldConfig > 0 || unpackOcclusion;
-        var unpackCollision = importCollision > 0;
+        var unpackAll = !importIntoExistingMap;
+        var unpackSounds = unpackAll || importMobys > 0;
+        var unpackAssets = unpackAll || importMobys > 0 || importSky > 0 || importCollision > 0 || importTfrags > 0 || importTies > 0 || importShrubs > 0;
+        var unpackGameplay = unpackAll || importMobys > 0 || importWorldConfig > 0 || importMisc > 0;
+        var unpackOcclusion = unpackAll || importTies > 0 || importTfrags > 0 || importMobys > 0;
+        var unpackWorldInstances = unpackAll || importTies > 0 || importShrubs > 0 || importWorldConfig > 0 || unpackOcclusion;
+        var unpackCollision = unpackAll || importCollision > 0;
 
         if (racVersion == RCVER.UYA)
             unpackGameplay |= unpackWorldInstances;
@@ -1376,7 +1387,7 @@ public class LevelImporterWindow : EditorWindow
     void ImportCode(string mapBinFolder, string mapResourcesFolder, GameRegion racRegion, List<PackerImporterWindow.PackerAssetImport> assetImports, GameObject rootGo)
     {
         var racVersion = ImportSourceRacVersion();
-        var binCodeFolder = Path.Combine(mapBinFolder, FolderNames.CodeFolder);
+        var binCodeFolder = Path.Combine(mapBinFolder, FolderNames.BinaryCodeFolder);
         var resourcesCodeFolder = Path.Combine(mapResourcesFolder, FolderNames.GetMapCodeFolder(racVersion, racRegion));
         if (Directory.Exists(resourcesCodeFolder)) Directory.Delete(resourcesCodeFolder, true);
         Directory.CreateDirectory(resourcesCodeFolder);
